# perf: multi thread file hashing and increase buffer size